Refactoring preprocessing
Mostly worked on pathway_network.py. Refactored it to fit new
dependencies and better coding practices.

Split binn into binn/model and binn/analysis

Maintained functionality.
