+++ binn/__init__.py
@@ -1,6 +1,6 @@
-from .binn import BINN
-from .logger import SuperLogger
-from .network import Network
-from .importance_network import ImportanceNetwork
-from .sklearn import BINNClassifier
-from .explainer import BINNExplainer
+from .model.binn import BINN
+from .model.logger import SuperLogger
+from .model.pathway_network import PathwayNetwork
+from .analysis.importance_network import ImportanceNetwork
+from .model.sklearn import BINNClassifier
+from .analysis.explainer import BINNExplainer
