+++ data/README.md
@@ -1,0 +1,2 @@
+translation --> uniprot_2_reactome_2025_01_14.txt
+pathways --> reactome_pathways_relation_2025_01_14.txt
