+++ setup.py
@@ -9,7 +9,7 @@ requirements = [
     "shap",
     "matplotlib",
     "plotly",
-    "nbformat>=4.2.0",
+    "nbformat",
     "kaleido",
 ]
